+++ resources/release-notes.md
@@ -8,5 +8,6 @@
 - Add definition of `mag.bart.Analysis` and other classes for analyzing Bartington data
 - Add definition of views for Bartington science, PSD and spectrogram
 - Add new processing classes for Bartington data
+- Add tests for Bartington and HelioSwarm analysis
 - Rename `mag.process.Filter` to `mag.process.EventFilter`
 - Fix IMAP export to CDF by correcting for UTC leap seconds